Resolved code smell and improved test coverage
