+++ packages/bng-metric-service/__tests__/bng-metric-extraction-service.spec.js
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import path from 'path'
-import { extractMetricContent } from '../src/service.js'
+import extractMetricContent from '../src/helpers/bng-metric-extraction-service.js'
 jest.mock('../src/service.js')
 
 describe('BNG data extrator service test', () => {
@@ -8,11 +8,13 @@ describe('BNG data extrator service test', () => {
   const currentPath = process.cwd()
 
   beforeEach(() => {
-    readableStream = fs.createReadStream(path.join(path.resolve(currentPath, 'packages', 'bng-metric-service', '__tests__/metricfiles/metric-file.xlsm')))
+    readableStream = fs.createReadStream(path.join(path.resolve(currentPath, 'packages/bng-metric-service/__mock-data__/metric-file', 'metric-file.xlsm')))
   })
 
-  it.skip('must extract all the configured excel sheets in a biodiversity metric file', async () => {
+  it('must extract all the configured excel sheets in a biodiversity metric file', async () => {
+    // const spy = jest.spyOn(bngMetricService, 'extractMetricContent').mockImplementation(contentInputStream => {})
     const response = await extractMetricContent(readableStream)
     expect(Object.keys(response).length).toBe(2)
+    // expect(spy).toHaveBeenCalled()
   })
 })
