+++ packages/bng-metric-service/src/service.js
@@ -1,14 +1,3 @@
-import { habitatBaselineExtractionConfig, startExtractionConfig } from './extractors/extractionconfig/configuration.js'
-import BngMetricSingleDataExtractor from './extractors/BngMetricSingleDataExtractor.js'
-
-const extractMetricContent = (contentInputStream) => {
-  const bngSingleExtractor = new BngMetricSingleDataExtractor()
-  const extractionConfiguration = {
-    startPage: startExtractionConfig,
-    siteHabitatBaseline: habitatBaselineExtractionConfig
-  }
-
-  return bngSingleExtractor.extractContent(contentInputStream, extractionConfiguration)
-}
+import extractMetricContent from './helpers/bng-metric-extraction-service.js'
 
 export default { extractMetricContent }
